[I compiled a **MUST READ** list with useful links regarding to rendering perf optimizations](https://github.com/trezor/trezor-suite/issues/317#issuecomment-528364361)

## Account menu 
_(rendering full acc menu on all all seed)_
- main problem: rerendering all accounts (even accs that were rendered before) on each change to the accounts array
- solution: `React.memo`!

### Before
We had **156 commits** (# of changes that React makes to the DOM) that took **more than 20ms**
<img width="444" alt="20msmefore" src="https://user-images.githubusercontent.com/6961901/64345753-2ec15180-cff1-11e9-9106-a0f43bf9bc01.png">

 **12 commits** that took **more than 80ms**
<img width="443" alt="80msbefore" src="https://user-images.githubusercontent.com/6961901/64346016-a8f1d600-cff1-11e9-9b9f-3a68f9281a00.png">


### After
 just **8 commits** above 20ms 
<img width="415" alt="20msafter" src="https://user-images.githubusercontent.com/6961901/64345755-2ec15180-cff1-11e9-82e6-9cca18c25e30.png">

**NONE above 80ms**